[LB-459] feat : 트레이싱 설정
