+++ build.gradle
@@ -46,6 +46,13 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+	implementation 'org.springframework.boot:spring-boot-starter-actuator'
+	runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
+
+	implementation 'io.micrometer:micrometer-tracing-bridge-otel' // 트레이싱
+	implementation 'io.opentelemetry:opentelemetry-exporter-otlp' // otel exporter
+	implementation 'net.ttddyy.observation:datasource-micrometer-spring-boot:1.1.2' // JDBC 트레이싱
 }
 
 tasks.named('test') {
