+++ src/main/java/com/lgcms/leveltest/config/KafkaConfig.java
@@ -1,0 +1,9 @@
+package com.lgcms.leveltest.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@Configuration
+@EnableKafka
+public class KafkaConfig {
+}
